api/requirements.txt
nomics-python==3.2.0
